removed compiler warnings
git-svn-id: http://svn.seqan.de/seqan/trunk@9024 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/splazers/razers_matepairs.h
@@ -569,7 +569,7 @@ void compactPairMatches(TMatches &matches, TCounts & /*cnts*/, RazerSOptions<TSp
 		if ((*it).orientation == '-') continue;
 		if (readNo == ((*it).rseqNo >> 1))
 		{ 
-			if ((*it).pairScore <= scoreDistCutOff) continue;
+			if ((int)(*it).pairScore <= scoreDistCutOff) continue;
 			if (++hitCount >= hitCountCutOff)
 			{
 #ifdef RAZERS_MASK_READS
